+++ db/migrate/20220209131448_create_variants.rb
@@ -1,7 +1,7 @@
 class CreateVariants < ActiveRecord::Migration[6.1]
   def change
     create_table :variants do |t|
-      t.integer :shopify_id
+      t.bigint :shopify_id
       t.string :title
       t.string :price
       t.string :sku
